refactor(cff): extract index module for CFF INDEX parsing/creation
