+++ oxidize-pdf-core/src/text/fonts/cff/mod.rs
@@ -3,4 +3,5 @@
 //! This module provides CFF parsing primitives extracted from the monolithic
 //! `cff_subsetter` module, intended to be shared across CFF-related components.
 
+pub mod index;
 pub mod types;
